requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry

azure-functions
#added for document intelligence
cryptography
azure-storage-blob
azure-eventhub
azure-identity
requests
openai
